Add rm_background.py utility file
The commit incorporates a new utility file named rm_background.py into the API. The utilities in this file are created to handle background removal functionalities. It checks if an icon has a transparent background and also contains the functionality to remove the background. The 'icon' and 'remove_bg' utilities are then imported into icon.py.