The provided commit message does not match the changes reflected in the code diffs. Based on the given changes, a more suitable commit message would be:
"Enable API response and reduce max attempts count"

The changes made in the code have enabled the response from HTTPClient request in the google_search.py file. This was previously commented out, and now the data, instead of being hardcoded, is being extracted directly from the response. The number of maximum attempts, in the icon.py file,